requirements.txt
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asyncpg==0.31.0
bcrypt==3.2.2
cffi==2.0.0
click==8.3.1
dlib==20.0.0
face-recognition==1.3.0
face_recognition_models==0.3.0
fastapi==0.128.0
fastapi-cache2==0.2.2
fastapi-limiter==0.1.6
greenlet==3.3.0
h11==0.16.0
idna==3.11
invoke==2.2.1
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
numpy==2.4.0
pandas==2.3.3
passlib==1.7.4
pendulum==3.1.0
pillow==12.0.0
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
pytz==2025.2
redis==4.6.0
setuptools==80.9.0
six==1.17.0
sqladmin==0.22.0
SQLAlchemy==2.0.45
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.40.0
WTForms==3.1.2
